Add TaskList Isolation Group Assignment
Update TaskListPartitionConfig with an explicit mapping of partition ids to metadata about the partition, rather than just a number of partitions. Store the isolation groups associated with each partition there.

Update TaskListStatus to include metrics about the TaskList with a per isolation group granularity. This is returned as part of DescribeTaskList, which will be used by the root partition to aggregate metrics from each of its children. It's currently used only to check if a child partition has finished draining.